requirements.txt
requests>=2.31.0
python-docx>=1.1.0
pydub>=0.25.1
openai-whisper>=20231117
nltk>=3.8.1
python-dotenv>=1.0.0
click>=8.1.7
tqdm>=4.66.1
colorama>=0.4.6
chardet>=5.0.0
